+++ src/templates/byTag.jsx
@@ -2,20 +2,21 @@ import React from "react";
 import Helmet from "react-helmet";
 import { graphql } from "gatsby";
 import Layout from "../layout";
-import PostListing from "../components/PostListing/PostListing";
+import DirectoryListing from "../components/PostListing/DirectoryPostListing";
 import config from "../../data/SiteConfig";
 
-export default class CategoryTemplate extends React.Component {
+export default class TagTemplate extends React.Component {
   render() {
-    const { category } = this.props.pageContext;
+    const { tag } = this.props.pageContext;
     const postEdges = this.props.data.allMarkdownRemark.edges;
     return (
       <Layout>
-        <div className="category-container">
-          <Helmet
-            title={`Posts in category "${category}" | ${config.siteTitle}`}
-          />
-          <PostListing postEdges={postEdges} />
+        <div className="tag-container">
+          <Helmet title={`Posts tagged as "${tag}" | ${config.siteTitle}`} />
+          <h1>{tag}</h1>
+          
+          <DirectoryListing postEdgesDirectory={postEdges} />
+        
         </div>
       </Layout>
     );
@@ -24,25 +25,22 @@ export default class CategoryTemplate extends React.Component {
 
 /* eslint no-undef: "off" */
 export const pageQuery = graphql`
-  query CategoryPage($category: String) {
+  query TagPage($tag: String) {
     allMarkdownRemark(
       limit: 1000
       sort: { fields: [fields___date], order: DESC }
-      filter: { frontmatter: { category: { eq: $category } } }
+      filter: { frontmatter: { tags: { in: [$tag] } } }
     ) {
       totalCount
       edges {
         node {
-          fields {
-            slug
-            date
-          }
-          excerpt
-          timeToRead
           frontmatter {
             title
+            website
+            twit
+            inst
+            category
             tags
-            date
           }
         }
       }
